feat(diffusers/pipelines): add pipelines like ControlNet, T2I-Adapter, UnCLIP, Pixart, AnimateDiff and etc. (#562)
* feat(diffusers/pipes): animatediff

* feat(diffusers/pipelines): add DeepFloyd IF

* feat(diffusers/pipes): add i2vgen_xl

* feat(diffusers/pipelines): add pixart-alpha

* feat(diffusers/pipelines): add shap_e

* feat(diffusers/pipelines): sd gligen

* feat(diffusers/pipelines): add t2i_adapter

* feat(diffusers/pipes): add UnCLIP

* feat(diffusers/pipelines): add ControlNets

* feat(diffusers/pipelines): add feats about loaders

* feat(diffusers/pipelines): add pipelines to init

* feat(diffusers/pipelines): add generator for vae encode

* fix(diffusers/scheduler): fix dtype cast in DDIM step

* refactor(diffusers/pipes): add ms.mutable to avoid repetitive compilation

+++ mindone/diffusers/schedulers/scheduling_ddim.py
@@ -445,7 +445,7 @@ class DDIMScheduler(SchedulerMixin, ConfigMixin):
         # 5. compute variance: "sigma_t(η)" -> see formula (16)
         # σ_t = sqrt((1 − α_t−1)/(1 − α_t)) * sqrt(1 − α_t/α_t−1)
         variance = self._get_variance(timestep, prev_timestep)
-        std_dev_t = (eta * variance ** (0.5)).to(dtype=prev_timestep.dtype)
+        std_dev_t = (eta * variance ** (0.5)).to(dtype)
 
         if use_clipped_model_output:
             # the pred_epsilon is always re-derived from the clipped x_0 in Glide
